src/backend/python/requirements.txt
requests
lxml
PyYAML
pandas
bottle
pm4py